Move ordman base view into separate files
