+++ PlattformOrdMan/UI/View/Base/ListDataType.cs
@@ -1,0 +1,11 @@
+﻿namespace PlattformOrdMan.UI.View.Base
+{
+    public enum ListDataType
+    {
+        DateTime,
+        Double,
+        Int32,
+        String,
+        Currency
+    }
+}
