+++ PlattformOrdMan/UI/View/Base/OrderManListView.cs
@@ -1,27 +1,14 @@
 ﻿using System;
 using System.Collections;
-using System.Data;
 using System.Drawing;
 using System.Windows.Forms;
-using System.Globalization;
-using System.ComponentModel;
-using Molmed.PlattformOrdMan.Data;
-using Molmed.PlattformOrdMan.UI.Controller;
+using Molmed.PlattformOrdMan.UI.View;
 
-namespace Molmed.PlattformOrdMan.UI.View
+namespace PlattformOrdMan.UI.View.Base
 {
     public partial class OrderManListView : ListView, IScrollableView
     {
-        public delegate void SortOrderSet(object sender, EventArgs e);
         public event SortOrderSet OnSortOrderSet;
-        public enum ListDataType
-        {
-            DateTime,
-            Double,
-            Int32,
-            String,
-            Currency
-        }
 
         protected const Int32 NO_COLUMN_INDEX = -1;
         private const Int16 WM_VSCROLL = 0x0115;
@@ -310,222 +297,5 @@ namespace Molmed.PlattformOrdMan.UI.View
                 OnSortOrderSet(this, new EventArgs());
             }
         }
-
-        private class ListViewComparerCurrency : ListViewComparerChiasma
-        {
-            private Int32 MySortColumnIndex;
-
-            public ListViewComparerCurrency(Int32 sortColumnIndex)
-                : base()
-            {
-                MySortColumnIndex = sortColumnIndex;
-            }
-
-            private bool GetPriceAndCurrency(String priceString, out String currencyString, out decimal price)
-            {
-                price = PlattformOrdManData.ParsePrice(priceString, out currencyString);
-
-                if (price == -1)
-                {
-                    return false;
-                }
-                else
-                {
-                    return true;
-                }
-            }
-
-            public override int Compare(Object object1, Object object2)
-            {
-                Boolean hasValue1, hasValue2;
-                int compareValue = 0;
-                String currencyString1 = "", currencyString2 = "", priceString1, priceString2;
-                decimal price1 = 0, price2 = 0;
-                ListViewItem listViewItem1, listViewItem2;
-
-                listViewItem1 = (ListViewItem)object1;
-                listViewItem2 = (ListViewItem)object2;
-                priceString1 = listViewItem1.SubItems[MySortColumnIndex].Text.Trim();
-                priceString2 = listViewItem2.SubItems[MySortColumnIndex].Text.Trim();
-                hasValue1 = GetPriceAndCurrency(priceString1, out currencyString1, out price1);
-                hasValue2 = GetPriceAndCurrency(priceString2, out currencyString2, out price2);
-                if (hasValue1 && hasValue2)
-                {
-                    if (String.Compare(currencyString1, currencyString2) == 0)
-                    {
-                        compareValue = price1.CompareTo(price2);
-                    }
-                    else
-                    {
-                        compareValue = String.Compare(currencyString1, currencyString2);
-                    }
-                }
-                else
-                {
-                    if (hasValue1)
-                    {
-                        compareValue = 1;
-                    }
-                    if (hasValue2)
-                    {
-                        compareValue = -1;
-                    }
-                    if (!hasValue1 && !hasValue2)
-                    {
-                        compareValue = 0;
-                    }
-                }
-                return compareValue * GetSortOrder();
-            }
-        }
-
-        private class ListViewComparerInt32 : ListViewComparerChiasma
-        {
-            private Int32 MySortColumnIndex;
-
-            public ListViewComparerInt32(Int32 sortColumnIndex)
-                : base()
-            {
-                MySortColumnIndex = sortColumnIndex;
-            }
-
-            public override int Compare(Object object1, Object object2)
-            {
-                Boolean hasValue1, hasValue2;
-                int compareValue = 0;
-                Int32 value1, value2;
-                ListViewItem listViewItem1, listViewItem2;
-
-                listViewItem1 = (ListViewItem)object1;
-                listViewItem2 = (ListViewItem)object2;
-                hasValue1 = Int32.TryParse(listViewItem1.SubItems[MySortColumnIndex].Text, out value1);
-                hasValue2 = Int32.TryParse(listViewItem2.SubItems[MySortColumnIndex].Text, out value2);
-                if (hasValue1 && hasValue2)
-                {
-                    compareValue = value1.CompareTo(value2);
-                }
-                else
-                {
-                    if (hasValue1)
-                    {
-                        compareValue = 1;
-                    }
-                    if (hasValue2)
-                    {
-                        compareValue = -1;
-                    }
-                    if (!hasValue1 && !hasValue2)
-                    {
-                        compareValue = 0;
-                    }
-                }
-                return compareValue * GetSortOrder();
-            }
-        }
-
-        private class ListViewComparerDouble : ListViewComparerChiasma
-        {
-            private Int32 MySortColumnIndex;
-
-            public ListViewComparerDouble(Int32 sortColumnIndex)
-                : base()
-            {
-                MySortColumnIndex = sortColumnIndex;
-            }
-
-            public override int Compare(Object object1, Object object2)
-            {
-                Boolean hasValue1, hasValue2;
-                int compareValue = 0;
-                ListViewItem listViewItem1, listViewItem2;
-                Double value1, value2;
-
-                listViewItem1 = (ListViewItem)object1;
-                listViewItem2 = (ListViewItem)object2;
-                hasValue1 = Double.TryParse(listViewItem1.SubItems[MySortColumnIndex].Text, out value1);
-                hasValue2 = Double.TryParse(listViewItem2.SubItems[MySortColumnIndex].Text, out value2);
-                if (hasValue1 && hasValue2)
-                {
-                    compareValue = value1.CompareTo(value2);
-                }
-                else
-                {
-                    if (hasValue1)
-                    {
-                        compareValue = 1;
-                    }
-                    if (hasValue2)
-                    {
-                        compareValue = -1;
-                    }
-                    if (!hasValue1 && !hasValue2)
-                    {
-                        compareValue = 0;
-                    }
-                }
-                return compareValue * GetSortOrder();
-            }
-        }
-
-        private class ListViewComparerString : ListViewComparerChiasma
-        {
-            private Int32 MySortColumnIndex;
-
-            public ListViewComparerString(Int32 sortColumnIndex)
-                : base()
-            {
-                MySortColumnIndex = sortColumnIndex;
-            }
-
-            public override int Compare(Object object1, Object object2)
-            {
-                int compareValue;
-                ListViewItem listViewItem1, listViewItem2;
-
-                listViewItem1 = (ListViewItem)object1;
-                listViewItem2 = (ListViewItem)object2;
-                compareValue = String.Compare(listViewItem1.SubItems[MySortColumnIndex].ToString(),
-                                           listViewItem2.SubItems[MySortColumnIndex].ToString(), true);
-                return compareValue * GetSortOrder();
-            }
-        }
-
-        public abstract class ListViewComparerChiasma : IComparer
-        {
-            private SortOrder MySortOrder;
-
-            public ListViewComparerChiasma()
-                : base()
-            {
-                MySortOrder = SortOrder.Ascending;
-            }
-
-            public abstract int Compare(Object object1, Object object2);
-
-            protected Int32 GetSortOrder()
-            {
-                if (MySortOrder == SortOrder.Ascending)
-                {
-                    return 1;
-                }
-                else
-                {
-                    return -1;
-                }
-            }
-
-            public virtual void SwitchSortOrder()
-            {
-                if (MySortOrder == SortOrder.Ascending)
-                {
-                    MySortOrder = SortOrder.Descending;
-                }
-                else
-                {
-                    MySortOrder = SortOrder.Ascending;
-                }
-            }
-        }
     }
-
 }
