+++ PlattformOrdMan/UI/View/Base/SortOrderSet.cs
@@ -1,0 +1,6 @@
+﻿using System;
+
+namespace PlattformOrdMan.UI.View.Base
+{
+    public delegate void SortOrderSet(object sender, EventArgs e);
+}
