+++ PlattformOrdMan/UI/View/Post/PostListView.cs
@@ -4,6 +4,7 @@ using System.ComponentModel;
 using Molmed.PlattformOrdMan.Data;
 using Molmed.PlattformOrdMan.UI.View;
 using PlattformOrdMan.Data.Conf;
+using PlattformOrdMan.UI.View.Base;
 
 namespace PlattformOrdMan.UI.View.Post
 {
